requirements-dev.txt
mypy >= 0.800
pytest
pytest-benchmark
pytest-cov
pytest-profiling
gprof2dot
epc
jedi
importmagic
